Change borrowing rate offset to 3%
