+++ test/local/variable-utilisation-rates-calculator.test.ts
@@ -15,35 +15,32 @@ const {expect} = chai;
 
 // From left: pool utilisation, expected Borrowing Rate, expected Deposit Rate
 const TEST_TABLE_RATES =  [
-  [0,	          0.05,	            0],
-  [0.01,	      0.0512,	        0.000511999],
-  [0.00000001,	  0.0500000012,	    0.0000000005],
-  [1e-9,	      0.05000000012,	5e-11],
-  [0.1,	          0.062,	        0.006199999],
-  [0.15,	      0.068,	        0.010199999],
-  [0.2,	          0.074,	        0.014799999],
-  [0.3,	          0.086,	        0.025799999],
-  [0.4,	          0.098,	        0.039199999],
-  [0.5,	          0.11,	            0.054999999],
-  [0.6,	          0.122,	        0.073199999],
-  [0.7,	          0.134,	        0.093799999],
-  [0.79,	      0.1448,	        0.114391999],
-  [0.799999,	  0.14599988,	    0.116799757],
-  [0.8,	          0.146,	        0.116799999],
-  [0.800001,	  0.14600302,	    0.116802561],
-  [0.81,	      0.1762,	        0.142721999],
-  [0.825,	      0.2215,	        0.182737499],
-  [0.85,	      0.297,	        0.252449999],
-  [0.875,	      0.3725,	        0.325937499],
-  [0.9,	          0.448,	        0.403199999],
-  [0.925,	      0.5235,	        0.484237499],
-  [0.95,	      0.599,	        0.569049999],
-  [0.975,	      0.6745,	        0.657637499],
-  [0.999,	      0.74698,	        0.746233019],
-  [1,	          0.75,	            0.749999999],
-  [1.1,	          0.75,	            0.749999999]
+  [0,             0.03,                 0],
+  [0.01,          0.0312,               0.000311999],
+  [0.00000001,    0.0300000012,         0.000000000300000012],
+  [0.1,           0.042,                0.004199999],
+  [0.15,          0.048,                0.007199999],
+  [0.2,           0.054,                0.010799999],
+  [0.3,           0.066,                0.019799999],
+  [0.4,           0.078,                0.031199999],
+  [0.5,           0.09,                 0.044999999],
+  [0.6,           0.102,                0.061199999],
+  [0.7,           0.114,                0.079799999],
+  [0.79,          0.1248,               0.098591999],
+  [0.8,           0.126,                0.100799999],
+  [0.799999,      0.12599988,           0.10079977700012],
+  [0.81,          0.1572,               0.127331999],
+  [0.800001,      0.12600312,           0.10080262100312],
+  [0.825,         0.204,                0.168299999],
+  [0.85,          0.282,                0.239699999],
+  [0.875,         0.36,                 0.314999999],
+  [0.9,           0.438,                0.394199999],
+  [0.925,         0.516,                0.477299999],
+  [0.95,          0.594,                0.564299999],
+  [0.975,         0.672,                0.655199999],
+  [0.999,         0.74688,              0.746133119],
+  [1,             0.75,                 0.749999999]
 ]
-
 // From left: totalLoans, totalDeposits, expected pool utilisation
 const POOL_UTILISATION =  [
   [1,	          10,	                    0.1],
